Fixed an error that occurs when arguments are added from the command line.

+++ trunk/Operators/TOF_SCD/IntegrateNorm.java
@@ -267,7 +267,7 @@ public class IntegrateNorm {
             BadEdgeWidth = Integer.parseInt( args[2].trim() );
          
          if( args.length >3 )
-            MaxCellEdge = Integer.parseInt( args[3].trim() );
+            MaxCellEdge = Float.parseFloat( args[3].trim() );
         
          }else
          {
